Create README.md
You are a Devops engineer and the organization you are working on needs to set up two configuration management server groups. One for Apache and another for Nginx. Being a Devops Engineer it is your task to deal with this configuration management issue.

Tasks to be performed in Ansible.
1. Create two server groups. One for Apache and one for Nginx.
2. Push two HTML files with their server information.

Make sure to start the services once the installation is done. Also, send post-installation messages for both server groups.
Ising Ansible Roles accomplish the above tasks.

Once Apache server configuration is done install Java on the server group using ansible role in a playbook.